requirements.txt
cachetools==4.2.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==5.1.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.8
